feat: slidable relationship_card  to perform actions (share, edit, re…

+++ lib/widgets/popup_menu_detail_relationship.dart
@@ -77,7 +77,10 @@ class _PopupMenuDetailRelationshipState
                     style: ButtonStyle(
                         backgroundColor: MaterialStatePropertyAll(Colors.red)),
                     onPressed: () {
-                      File(widget.user.imageUrl!).delete();
+                      if (widget.user.imageUrl! != '') {
+                        File(widget.user.imageUrl!).delete();
+                      }
+
                       context.read<UsReListBloc>().add(
                           DeleteUsRe(usReId: widget.userRelationship.usReId!));
                       context
